filters.json: 2018-10-02 edition of 23 & 25 'Sponsored (Experimental)' parts A & B
Remove the 'Paid' selector as it triggers on such posts forwarded
(shared) by a friend, which may contain comments or discussion.  If we
had a way to express 'only try this filter on posts NOT from a friend or
a Page which I follow', it would be fine.  Some day.

Protect 'part B' against Pages the user is following.  The filter now
means 'if the post is from a Page _which I do not follow_, and Post
Action says someone likes something' -- i.e. a non-subscribed Page post
which FB has inserted with the excuse that my friends like that Page.