filters.json: tell users of 6 old Sponsored filters to upgrade (lite)
This only affects the titles of the filters and the appearance of the
hider messages, so not everyone will see it.  Obviously anyone who looks
in their filter table will see it.  Other users will see it only if (1)
at least one of the filters is still catching posts; (2) they haven't
done 'Convert to local' on that filter; (3) they haven't edited the
actions of that filter.  If all that holds, they'll see a note in their
'post hidden, click to unhide' buttons.

I don't actually know if the old filters are still effective for
*anyone*, but there is fairly compelling circumstantial evidence: a
continuing trickle of people who join the group to ask why their
Sponsored filters 'just broke'; as if FB are still slowly rolling the
changes out.

I have a more aggressive set of changes(*) which actually pop up the
Control Panel for everyone who has the old filters installed.  But this
may be needlessly aggressive.  I'll put this out and see how it plays
for a few days or weeks.  We will of course get support questions from
people who notice the new text and can't figure out that when it says to
do A-B-C, the thing to do is A-B-C.  That will tell us that the title &
hider changes are effective.  If we *also* continue to see questions
from people who need to change to the v24 Sponsored filter and *didn't*
notice thos changes, then I can push out the more aggressive changes.

The 6 obsolete Sponsored filters are tweaked in 5 ways:

    - custom notes:
        . prefix with '[PLEASE SEE http://tiny.cc/sfx-v24-spfilter] '
        . reduce lengths so they don't wrap
        . add CSS (orange background) to make them striking
    - titles:
        . prefix with '[OBSOLETE: PLEASE SEE http://tiny.cc/sfx-v24-spfilter] '
        . edit a couple others: 'OLD' -> 'ANCIENT', main -> 'OLD-MAIN'

Affected filters:

     2 'Hide Sponsored/Suggested Posts (2019-02-03)' (renamed 'OLD-MAIN')
    23 'Sponsored/Suggested Posts (Experimental 2018-11-08 part A)'
    24 'Hide Sponsored/Suggested Posts (OLD)' (renamed 'ANCIENT')
    25 'Sponsored/Suggested Posts (Experimental 2018-11-08 part B)'
    26 'Sponsored/Suggested Posts (Experimental part D, 2019-04-05)'
    27 'Sponsored/Suggested Posts (Experimental part C, 2019-03-05)'

[[[

(*) The more aggressive changes are like so:

Add to the 'actions' of the 6 obsolete filters:

		}, { "action": "copy-to-tab", "tab": "  ! PLEASE CHECK"
		}, { "action": "copy-to-tab", "tab": "  ! YOUR SOCIAL"
		}, { "action": "copy-to-tab", "tab":  " ! FIXER FILTERS!!"

(spacing is intentional to sort in the correct order)

Add to the 'rules' of the 3 filters with 'OR' actions (2, 23, 24):

		}, { "target": "any", "operator": "contains_selector", "condition": { "text": "[sfx_post='9']>div" }

This causes these filters to catch the 9th post in the feed, thus
instantiating the 3 message tabs.  9th chosen to avoid any problems with
startup timing; it should be in the 2nd tranche of posts sent by the FB
server, long after everything is fully up and running.  The 3 'AND'
filters cannot be modified this way (no full boolean logic).

Holding these changes in reserve for now.

]]]