hideable.json: add 350 'Above Group Feed: Introducing Messenger Rooms for Groups'
This selector might not be specific.  I've seen the targeted area used
for 2 different things; the other was today, about a video about racial
injustice.  This selector distinguishes between those two, but might
pick up any other top-of-news-feed splash they add which lacks an href
(the video had an href).  The 3 newest-style classes I'm selecting on
are the ones that seem most specific to *this* presentation, thus least
likely to be repeated on some other use of the area.

I'm not using the hider myself, instead have a dashed border CSS to
keep an eye on it.  If I see something else highlighted I'll be able to
either fix this, or abolish it if I can't make a sufficiently specific
hider.