hideable.json: fix Sponsored item hiders for real
2021010501 'Right Rail: Sponsored item' is updated to the prior contents of
2021012401 'Right Rail: Sponsored item (2)', except no '[role=button]'.

Turns out the 2nd for is how FB present ads to some people, then when the
ad is moused over, it morphs into the first form.  This means they can't
hide the 2nd form because while trying to hide it, it morphs into the 1st
form!  So, need a hider which encompasses both forms.

Unfortunately this makes the entire hider a 'parent' type hider, which is
slower, so now you see the ads briefly on every page load.