hideable.json: fix 40 LeftRail items
Change '[data-pagelet=page] [role=navigation]' to '[data-pagelet=LeftRail]'
in all 40 LeftRail items.  The 'Left Rail: Marketplace' item was hiding
almost the entire user interface on fb.com/marketplace/.  This fixes it,
and I can see that it is the more correct phrasing for this entire area.
Marketplace wasn't the first thing that had this problem (I forget what
was).

Affects hider IDs 2021010401..2021010436, 2021010601, 2021011001..2021011003