hideable.json: add 'Groups Feed: ad to join groups'
The selector for this is particularly ugly, derived from a single
example (which I have a copy of) from Ieda.  As with all these
new-layout selectors, it's my guess of what unique page-construction
details lead down to this item; the only definitive bit is
a[href*='groups/discover'], but that could hit all sorts of other
things, so then I construct a path to the box out of whatever
weird things it has -- in this case including 'flex-grow:0' and
'.t7l9tvuc:last-of-type{margin-right:0}'.  Which makes it fragile to
future FB changes, but unlikely to hide things I don't intend...