hideable.json: add and update some top bar buttons
- update 2020120706 'Header: 'My profile' button' to position hider correctly
- add 2021042901 'Header: 'Notifications' menu'
- add 2021042902 'Header: 'Create / Menu' menu'
- add 2021042903 'Header: 'Messenger' menu'
- add 2021042904 'Header: 'Pages' button'
- add 2021042905 'Header: 'Account' button [DISABLED]' (*)

(*) Support team agreed this was too dangerous to offer; this records
    its method, in case there's a future reason to offer it.

Some of these hiders use aria-label matching, which means they only
work when the FB user interface language is English.  For Create/Menu,
Messenger, and Account (DISABLED), I was unable to create a working
hider which didn't rely on aria-label.  For Pages and Notifications, I
supply aria-label as a 'backup' so the hider might continue to work
under some potential FB changes.  Also, only the aria-label
Notifications hider works on the persistent notifications page
(fb.com/notificaitons), as it doesn't provide a link to itself.