tweaks.json: add missing doc for 3 'Hide Notification Bubble' change
The '-mode' fiddling is to only catch *notifications* while avoiding
realtime indications that something you just did was received.  e.g.
when you approve a post to be posted or a member to join a group, or
remove a notification from the notifications menu, a popup tells you
that it went through.  These 'action' notifications are shown in the
opposite of your regular 'theme' (dark / light), while regular popup
notifications are in the normal theme.

There's some demand to also hide the 'action' notifications -- but I
consider some of them important UI elements.  In some cases they are
the only way offered to undo a potentially destructive action.  Some
others are trash worthy of hiding; but as we lack a :contains() type
operator, we cannot separate the wheat from the chaff.