tweaks.json: add 2021093001 'Fix date-stamp blobs (TEMP FIX)'
A stopgap fix for the fact that FB no longer notice our pointerout /
mouseout events, apparently insisting on isTrusted.  Thus, the screen
is littered with date stamp blobs.

The crude fix is to fade out such blobs after 5s.  Not shorter, because
this affects a whole class of such blobs, including the list of who
'liked' a post when hovering the blue thumb-with-count, and, I think,
some immediate action undo buttons.

This needs a real fix by SFx either no longer doing the hover thing to
provoke the permalink URL, or figuring out how to convince it we've
moused off.