hideable.json: fix old-FF compatibility of 2021041101, 2021041102, 2021082201, 2021101501, 2022021801, 2022022501
In commit a129bc24 I wrote: 'I think this anchor will be more stable.  But can't really test it, so will wait for user reports...'

fb.com/741404923932054 is a user report that Hide/Show doesn't work on FF 79.  I verified in FF 78esr: selectors of the form ':not(.this,.that)' do not work.  So, fix those selectors to be back-compatible: ':not(.this):not(.that)'.

Affects hiders: 'News Feed:{ Remember Password prompt, Take A Survey prompt, Add a WhatsApp Button, Finish Donation, Keep your Page up to date, Facebook ad discount}'