hideable.json: adapt to lack of [role=article]
Previously, each post was [role=article]; and so were each comment.  So
[role=article] could catch either, while [role=article] [role=article]
targeted comments specifically.

Now, the top of a post reliably has [aria-posinset], but never
[role=article].  Replace those references, except in a couple of ancient
hiders which are probably irrelevant and I have no way to even try to
check.

A few hiders *are* targeted at comment elements, so they still use
[aria-posinset] [role=article].