hideable.json: add 2022102600 "Post: 'view product' widget" hideable.json: add 2022102601 "Post: 'Click to view products' button and floaters"
The 'Click to view products' button hider appears completely out of
place, in fact breaking the page layout.  But as long as you ignore the
odd appearance and click on it, it does the right thing.

The 2nd hider combines two hiders which are completely unrelated at the
page HTML level.  The 2nd item is ephemeral and extremely difficult to
hide with Hide/Show without some browser console skullduggery.

hideable.json: 20220823{18,25,34,39,40,47,48,49} adjust gib names to match code changes