filters.json: 2021082601 'Suggested' add group name to the click-to-see filters.json: 2021082601 'Suggested' fix totally wrong negative lookahead
The negative lookahead is meant to capture group names in 'More like'
and 'Suggested groups' fakeposts, which are structured with a 'xyz
number of posts a day/week/month' teaser followed by actual group names.
Which is then, these days, followed by Facebook's selfarrhea, so use a
negative lookahead to chop that off.