hideable.json: add 2023071201 'Groups Feed: Chats you should join'
This was tricky because the blob-to-hide moves between top and right
columns depending on zoom level.  This is not done with pure CSS;
rather, FB rewrite *part* of the DOM on width changes.  I had to find
a node which covered the entire blob but was not rewritten during a
width change -- otherwise the blob would reappear on zoom in/out past
the boundary size (but then disappear again on reload).

In the 2-column 'zoomed out' presentation, this also has the benefit
of removing the space used, allowing the center column to expand.