hideable.json: add 2023092501 'Media 1' hideable.json: add 2023092502 'Media 2' hideable.json: add 2023092503 'Media 3'
Names are short because the display areas for the hider-shaders are small.
Even with these short names the prompts are truncated (but readable).

They're positional.  Different screens (like one's timeline vs. Top Stories
news feed vs. a Group) have different collections of media buttons.  But
it's OK to hide them all, since if you click in the text box they pop up an
editor which offers all of those options and many more.