Try to fix overly eager hiding of shared Group posts
filters.json: 2021082601 'Suggested' remove the 'Follow|Join' clauses
filters.json: 2021082601 'Suggested' add a clause looking for 'is a member'

per HTML collected at fb.com/1129243368481539.  The removed clauses catch posts
which were legitimately shared by, posted by, or commented by a friend.  The
intent is to catch only posts which FB have shoved into a feed from some random
(public) group one is not a member of, just because a friend is in the group.